Statistics in Leisure and Hospitality
Hiring and opening statistics in Leisure and Hospitality. Job opportunities in this field depend on the month you are applying, the peak is in May. The job opportunities increased by 3 fold in 10 years. From job opportunity point of view, the Leisure and Hospitality has a better odds compare to Education and Health Services.